:truck: Move routes to shared
